pyproject.toml
[project]
name = "oversight-curriculum"
version = "1.0.0"
description = "AI Safety & Reasoning System - Oversight Curriculum"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Oversight Curriculum Team", email = "team@example.com" }
]
keywords = ["ai-safety", "reasoning", "oversight", "curriculum", "claude", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.8"
dependencies = [
    "anthropic>=0.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "httpx>=0.24.0",
    "tenacity>=8.2.0",
    "structlog>=23.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/oversight-curriculum/oversight-curriculum"
Repository = "https://github.com/oversight-curriculum/oversight-curriculum"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["oversight*"]
